xiaohe-llmops-api/pyproject.toml
[project]
name = "xiaohe-llmops-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "cos-python-sdk-v5>=1.9.38",
    "faiss-cpu>=1.12.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-login>=0.6.3",
    "flask-migrate>=4.1.0",
    "flask-wtf>=1.2.2",
    "injector>=0.22.0",
    "jieba>=0.42.1",
    "jinja2>=3.1.6",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain-ollama>=0.3.8",
    "langchain-openai>=0.3.33",
    "langchain-weaviate>=0.0.5",
    "langgraph>=0.6.7",
    "markdown>=3.9",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.9",
    "pyjwt>=2.10.1",
    "pypdf>=6.0.0",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "tiktoken>=0.11.0",
    "unstructured>=0.18.15",
    "weaviate-client>=4.16.10",
    "werkzeug>=3.1.3",
    "wtforms>=3.2.1",
]
